애플리케이션 구동시 초기 데이터

+++ src/main/java/com/blogdev/posts/dto/RqPostsListDto.java
@@ -1,4 +1,4 @@
 package com.blogdev.posts.dto;
 
-public record RqPostsListDto (int id,String title){
+public record RqPostsListDto (Long id,String title){
 }
